Feature/sabi 115 (#129)
fixes #115 

* captcha patch management
* Testdata and DDLs
* Typos
* Prepeared Plaguelisting API+Impl in Backend
* WebApp API Gateway
* DRY Refactorings
* UseCase: List active Plagues in PlageCenter finished.
* UseCase: List past Plagues in PlageCenter finished.
* Adding CodeQL badge to readme.md
* Usecase: Add new Plague records (rest and service layer) - UI refresh is open.
* SDK: Patchmanagement
* UseCase: Adding PLague Records finished

+++ sabi-server/src/main/java/de/bluewhale/sabi/persistence/model/PlagueRecordEntity.java
@@ -42,11 +42,11 @@ public class PlagueRecordEntity extends Auditable {
 
     @Column(name = "plague_id", nullable = false, insertable = true, updatable = true, length = 20, precision = 0)
     @Basic
-    private Long plagueId;
+    private Integer plagueId;
 
     @Column(name = "observed_plague_status", nullable = false, insertable = true, updatable = true, length = 20, precision = 0)
     @Basic
-    private Long observedPlagueStatus;
+    private Integer observedPlagueStatus;
 
     /**
      * The plague intervall id groups all records together, that belong to the same
